requirements.txt
asgiref==3.8.1
Django==4.2.17
djangorestframework==3.16.0
exceptiongroup==1.3.0
iniconfig==2.1.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.1.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
